refactor: #161 - useNotification 훅 정의 및 비즈니스 로직 분리 % 유틸 함수 추가
