+++ src/components/NotificationViewer/index.tsx
@@ -1,19 +1,16 @@
-import { MouseEvent, useEffect, useRef, useState } from 'react';
-import { useNavigate } from 'react-router-dom';
+import { MouseEvent } from 'react';
 import styled from '@emotion/styled';
 import { Notification } from '@type';
 import Icon from '@components/Icon';
 import Spinner from '@components/Spinner';
-import {
-  useFetchGetNotifications,
-  useFetchReadNotifications,
-} from '@apis/notifications';
 import { ANGOLA_STYLES } from '@styles/commonStyles';
 import {
   CommentNotificationItem,
   FollowNotificationItem,
   LikeNotificationItem,
 } from './components';
+import useNotification from './hooks/useNotification';
+import { filterSameFollowerNotifications } from './utils';
 
 const decideNotificationType = ({
   notification,
@@ -33,97 +30,23 @@ const decideNotificationType = ({
   }
 };
 
-interface NotificationViewePropsType {
+interface NotificationViewerPropsType {
   handleClickCloseViewer: VoidFunction;
 }
 
 const NotificationViewer = ({
   handleClickCloseViewer,
-}: NotificationViewePropsType) => {
-  const navigate = useNavigate();
-  const viewerBackGroundRef = useRef(null);
-
-  useEffect(() => {
-    // Viewer 바깥 영역 클릭 시, 뷰어 닫아지도록
-    const handleClickViewerBackGround = (e: Event) => {
-      if (
-        viewerBackGroundRef.current &&
-        e.target === viewerBackGroundRef.current
-      ) {
-        handleClickCloseViewer();
-      }
-    };
-
-    window.addEventListener('click', handleClickViewerBackGround);
-
-    return () => {
-      window.removeEventListener('click', handleClickViewerBackGround);
-    };
-  });
-
-  // 모든 알림 볼 건지 or 새로운 알림만 볼 건지
-  const [isShowAllNotifications, setIsShowAllNotifications] = useState(false);
-
-  // 알림 목록 조회 data
+}: NotificationViewerPropsType) => {
   const {
+    viewerBackGroundRef,
+    isShowAllNotifications,
+    setIsShowAllNotifications,
     getNotificationsData,
     isGetNotificationsLoading,
-    getNotificationRefetch,
-  } = useFetchGetNotifications();
-
-  // 알림 읽음 처리 API
-  const { readNotificationsMutate, isReadNotificationsLoading } =
-    useFetchReadNotifications();
-
-  const handleClickReadNotifications = () => {
-    readNotificationsMutate(undefined, {
-      onSuccess: () => {
-        getNotificationRefetch();
-      },
-    }); // TODO:MinwooP - 에러 처리
-  };
-
-  // notification Item 클릭 시 페이지 이동 로직
-  const handleClickListItem = (
-    e: MouseEvent<HTMLLIElement>,
-    notification: Notification,
-  ) => {
-    if (e.target instanceof HTMLSpanElement && e.target.id === 'userName') {
-      // Item내 userName 클릭 시 해당 유저페이지로 이동
-      navigate(`/user/${notification.author._id}`);
-    } else {
-      if (notification.post === null) {
-        navigate(`/user/${notification.author._id}`); // 팔로우 알림 클릭 시 => 작성자 유저 페이지로 이동
-      } else {
-        // 좋아요, 댓글 알림 클릭 시 시 => 해당 포스트 페이지로 이동
-        navigate(`/post/${notification.post}`);
-      }
-    }
-    handleClickCloseViewer();
-  };
-
-  const filterSameFollowerNotifications = (
-    notificationList: Notification[],
-  ) => {
-    const follwerList: string[] | null = [];
-    const filteredNotificationList: Notification[] | null = [];
-
-    notificationList.forEach((notification) => {
-      if (notification.post === null) {
-        // 팔로우 알림일 때
-        if (!follwerList.includes(notification.author.fullName)) {
-          // follweList에 없을 때 추가
-          follwerList.push(notification.author.fullName);
-          filteredNotificationList.push(notification);
-        }
-      } else {
-        // 좋아요, 댓글 알림일 시 바로 List에 추가
-        filteredNotificationList.push(notification);
-      }
-    });
-
-    return filteredNotificationList;
-  };
+    isReadNotificationsLoading,
+    handleClickReadNotifications,
+    handleClickListItem,
+  } = useNotification({ handleClickCloseViewer });
 
   const showAllNotificationList = () => {
     if (getNotificationsData?.length === 0) {
@@ -234,7 +157,7 @@ const NotificationViewerBackGround = styled.div`
   background-color: rgba(0, 0, 0, 0);
   width: 100%;
   height: 100vh;
-  z-index: 200; // TODO: 나중에 z-index 기준 컴포넌트 별로 설정 필요
+  z-index: 20;
   position: fixed;
   top: 0;
   left: 0;
@@ -323,6 +246,7 @@ const ReadNotificationButton = styled.button`
   background: ${ANGOLA_STYLES.color.white};
   box-shadow: ${ANGOLA_STYLES.shadow.button.default};
   cursor: pointer;
+
   &:hover {
     box-shadow: ${ANGOLA_STYLES.shadow.button.hover};
   }
