+++ src/components/NotificationViewer/utils/index.ts
@@ -1,0 +1,21 @@
+import { Notification } from '@type/index';
+
+export const filterSameFollowerNotifications = (
+  notificationList: Notification[],
+) => {
+  const follwerList: string[] | null = [];
+  const filteredNotificationList: Notification[] | null = [];
+
+  notificationList.forEach((notification) => {
+    if (notification.post === null) {
+      if (!follwerList.includes(notification.author.fullName)) {
+        follwerList.push(notification.author.fullName);
+        filteredNotificationList.push(notification);
+      }
+    } else {
+      filteredNotificationList.push(notification);
+    }
+  });
+
+  return filteredNotificationList;
+};
